+++ Changelog.md
@@ -1,3 +1,4 @@
 ## 2.1.0
 
 - Version numbering changed from "x.x.x.x" to "x.x.x". Introduced readme and changelog files.
+- Partially fixed 0ADD (spawn car like a cheat) for GTA3. Now opcode can spawn not only the tank. Still no support for models with id greater than 127.